+++ pyproject.toml
@@ -51,7 +51,6 @@ dependencies = [
     "pycountry==22.3.5",
     "pyomo==6.6.1",
     "pypsa==0.30.2",
-    "pytest>=8.3.5",
     "pyyaml>=6.0.2",
     "rasterio==1.3.8",
     "scipy==1.11.3",
@@ -82,6 +81,7 @@ dev = [
     "sphinx",
     "sphinx-book-theme",
     "myst-parser",
+    "pytest",
 ]
 
 [tool.setuptools.package-dir]
@@ -147,3 +147,8 @@ quote-style = "double"
 indent-style = "space"
 skip-magic-trailing-comma = false
 line-ending = "auto"
+
+[tool.pytest.ini_options]
+testpaths = [
+    "workflow/scripts/test"
+]
