+++ workflow/scripts/test/conftest.py
@@ -6,7 +6,6 @@ These fixtures provide reusable test data and network configurations.
 """
 import os
 import sys
-import tempfile
 
 import numpy as np
 import pandas as pd
@@ -301,251 +300,3 @@ def base_network():
         co2_emissions=0,
     )
     return n
-
-
-@pytest.fixture
-def land_use_network(base_network):
-    """
-    Adapt base network for land use constraint testing.
-
-    Extends the base network with specific parameters relevant to land use constraints.
-    """
-    n = base_network.copy()
-
-    # Add another generator in region_a to test constraints with multiple generators in the same region
-    n.add(
-        "Generator",
-        "wind3",
-        bus="z1",
-        p_nom=0,
-        p_nom_extendable=True,
-        carrier="onwind",
-        capital_cost=1050,
-        marginal_cost=0.11,
-        p_max_pu=pd.Series(0.85, index=n.snapshots),
-        land_region="region_a",
-        p_nom_max=300,
-    )
-
-    return n
-
-
-@pytest.fixture
-def reserve_margin_network(base_network):
-    """
-    Adapt base network for ERM and PRM constraint testing.
-
-    Extends the base network with parameters relevant to reserve margin testing.
-    """
-    n = base_network.copy()
-
-    # Create a higher peak in load profile for reserve margin testing
-    # Make load profile peaky with peak at hour 11 for region 1 and hour 18 for region 2
-    load_profile_region1 = pd.Series(
-        np.concatenate([np.linspace(800, 1200, 12), np.linspace(1200, 800, 12)]),
-        index=n.snapshots,
-    )
-
-    load_profile_region2 = pd.Series(
-        np.concatenate([np.linspace(500, 700, 18), np.linspace(700, 500, 6)]),
-        index=n.snapshots,
-    )
-
-    # Update load profiles
-    n.loads_t.p_set.loc[:, "load1"] = load_profile_region1
-    n.loads_t.p_set.loc[:, "load2"] = load_profile_region1 * 0.75
-    n.loads_t.p_set.loc[:, "load3"] = load_profile_region2
-
-    return n
-
-
-@pytest.fixture
-def policy_network(base_network):
-    """
-    Adapt base network for policy constraint testing (RPS, TCT).
-
-    Extends the base network with parameters needed for policy constraint testing.
-    """
-    n = base_network.copy()
-
-    # For policy constraints, we want clearer isolation of regions
-    # Add a nuclear generator (non-renewable but clean)
-    n.add(
-        "Generator",
-        "nuclear1",
-        bus="z3",
-        p_nom=300,
-        p_nom_extendable=True,
-        carrier="nuclear",
-        capital_cost=3000,
-        marginal_cost=5,
-        p_max_pu=pd.Series(1.0, index=n.snapshots),
-        p_nom_max=1500,
-    )
-
-    return n
-
-
-@pytest.fixture
-def rps_config():
-    """Create a config dictionary and temporary files for RPS constraints."""
-    # Create temporary portfolio standards file
-    portfolio_standards_file = tempfile.NamedTemporaryFile(delete=False, suffix=".csv")
-    pd.DataFrame(
-        [
-            {
-                "name": "CA_RPS",
-                "region": "CA",
-                "carrier": "solar, onwind",
-                "planning_horizon": "2030",
-                "pct": 0.5,
-            },
-            {
-                "name": "TX_RPS",
-                "region": "TX",
-                "carrier": "solar, onwind",
-                "planning_horizon": "2030",
-                "pct": 0.3,
-            },
-        ],
-    ).to_csv(portfolio_standards_file.name, index=False)
-
-    # Create temporary RPS REEDS file
-    rps_reeds_file = tempfile.NamedTemporaryFile(delete=False, suffix=".csv")
-    pd.DataFrame(
-        [
-            {"st": "CA", "t": "2030", "rps_all": 0.6},
-            {"st": "TX", "t": "2030", "rps_all": 0.4},
-        ],
-    ).to_csv(rps_reeds_file.name, index=False)
-
-    # Create temporary CES REEDS file
-    ces_reeds_file = tempfile.NamedTemporaryFile(delete=False, suffix=".csv")
-    pd.DataFrame(
-        [
-            {"st": "CA", "t": "2030", "rps_all": 0.7},
-            {"st": "TX", "t": "2030", "rps_all": 0.5},
-        ],
-    ).to_csv(ces_reeds_file.name, index=False)
-
-    # Create config dictionary
-    config = {
-        "electricity": {
-            "portfolio_standards": portfolio_standards_file.name,
-        },
-    }
-
-    # Create a mock snakemake object
-    class MockSnakemake:
-        def __init__(self):
-            self.input = type(
-                "obj",
-                (object,),
-                {
-                    "rps_reeds": rps_reeds_file.name,
-                    "ces_reeds": ces_reeds_file.name,
-                },
-            )
-            self.params = type(
-                "obj",
-                (object,),
-                {
-                    "planning_horizons": ["2030"],
-                },
-            )
-
-    snakemake = MockSnakemake()
-
-    yield config, snakemake
-
-    # Clean up temporary files
-    os.unlink(portfolio_standards_file.name)
-    os.unlink(rps_reeds_file.name)
-    os.unlink(ces_reeds_file.name)
-
-
-@pytest.fixture
-def erm_config():
-    """Create a config dictionary with ERM settings."""
-    # Create a temporary PRM file
-    pd.DataFrame(
-        {
-            "name": ["test_region"],
-            "region": ["all"],
-            "prm": [0.15],  # 15% reserve margin
-            "planning_horizon": ["2030"],
-        },
-    ).to_csv("/tmp/test_prm.csv")
-
-    return {
-        "electricity": {
-            "SAFE_regional_reservemargins": "/tmp/test_prm.csv",
-        },
-    }
-
-
-@pytest.fixture
-def erm_multi_region_config():
-    """Create a config dictionary with ERM settings for multiple regions and different reserve margins."""
-    # Create a temporary PRM file with both regions
-    pd.DataFrame(
-        {
-            "name": ["west_erm", "east_erm"],
-            "region": ["west", "east"],
-            "prm": [0.15, 0.12],  # Different reserve margins
-            "planning_horizon": ["2030", "2030"],
-        },
-    ).to_csv("/tmp/multi_region_prm.csv")
-
-    return {
-        "electricity": {
-            "SAFE_regional_reservemargins": "/tmp/multi_region_prm.csv",
-        },
-    }
-
-
-@pytest.fixture
-def tct_config():
-    """Create a config dictionary and temporary files for TCT constraints."""
-    # Create temporary technology capacity targets file
-    tct_file = tempfile.NamedTemporaryFile(delete=False, suffix=".csv")
-    pd.DataFrame(
-        [
-            {
-                "name": "wind_west_min",
-                "region": "west",
-                "carrier": "onwind",
-                "planning_horizon": "2030",
-                "min": 300,
-                "max": float("nan"),
-            },
-            {
-                "name": "wind_west_max",
-                "region": "west",
-                "carrier": "onwind",
-                "planning_horizon": "2030",
-                "min": float("nan"),
-                "max": 500,
-            },
-            {
-                "name": "battery_west_min",
-                "region": "west",
-                "carrier": "battery",
-                "planning_horizon": "2030",
-                "min": 100,
-                "max": float("nan"),
-            },
-        ],
-    ).to_csv(tct_file.name, index=False)
-
-    # Create config dictionary
-    config = {
-        "electricity": {
-            "technology_capacity_targets": tct_file.name,
-        },
-    }
-
-    yield config
-
-    # Clean up temporary file
-    os.unlink(tct_file.name)
