+++ workflow/scripts/test/fixtures/ces_reeds.csv
@@ -1,0 +1,3 @@
+st,t,rps_all
+CA,2030,0.7
+TX,2030,0.5
