+++ workflow/scripts/test/fixtures/multi_region_prm.csv
@@ -1,0 +1,3 @@
+name,region,prm,planning_horizon
+west_erm,west,0.15,2030
+east_erm,east,0.12,2030
