+++ workflow/scripts/test/fixtures/rps_reeds.csv
@@ -1,0 +1,3 @@
+st,t,rps_all
+CA,2030,0.6
+TX,2030,0.4
