+++ workflow/scripts/test/fixtures/test_prm.csv
@@ -1,0 +1,2 @@
+name,region,prm,planning_horizon
+test_region,all,0.15,2030
